+++ src/main/java/com/lmax/disruptor/util/ThreadHints.java
@@ -15,12 +15,6 @@
 
 package com.lmax.disruptor.util;
 
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
-
-
-import static java.lang.invoke.MethodType.methodType;
-
 /**
  * This class captures possible hints that may be used by some
  * runtimes to improve code performance. It is intended to capture hinting
@@ -30,23 +24,6 @@ import static java.lang.invoke.MethodType.methodType;
  */
 public final class ThreadHints
 {
-    private static final MethodHandle ON_SPIN_WAIT_METHOD_HANDLE;
-
-    static
-    {
-        final MethodHandles.Lookup lookup = MethodHandles.lookup();
-
-        MethodHandle methodHandle = null;
-        try
-        {
-            methodHandle = lookup.findStatic(Thread.class, "onSpinWait", methodType(void.class));
-        }
-        catch (final Exception ignore)
-        {
-        }
-
-        ON_SPIN_WAIT_METHOD_HANDLE = methodHandle;
-    }
 
     private ThreadHints()
     {
@@ -58,20 +35,13 @@ public final class ThreadHints
      * invoking this method within each iteration of a spin-wait loop construct,
      * the calling thread indicates to the runtime that it is busy-waiting. The runtime
      * may take action to improve the performance of invoking spin-wait loop constructions.
+     *
+     * @deprecated Use Thread.onSpinWait() directly. This method previously existed to accomodate
+     * Java versions which do not have Thread.onSpinWait().
      */
+    @Deprecated
     public static void onSpinWait()
     {
-        // Call java.lang.Thread.onSpinWait() on Java SE versions that support it. Do nothing otherwise.
-        // This should optimize away to either nothing or to an inlining of java.lang.Thread.onSpinWait()
-        if (null != ON_SPIN_WAIT_METHOD_HANDLE)
-        {
-            try
-            {
-                ON_SPIN_WAIT_METHOD_HANDLE.invokeExact();
-            }
-            catch (final Throwable ignore)
-            {
-            }
-        }
+        Thread.onSpinWait();
     }
 }