+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - Minimum Java version now 11
     - Removed OSGI bundle fields from `MANIFEST.MF`
     - Issue #323 - `WorkerPool` and `WorkProcessor` have been removed, no more `Disruptor::handleEventsWithWorkerPool`
+    - Deprecated ThreadHints.onSpinWait()
 
 ## 3.4.2